Move Sigma tests into the Module tests.
